Fixed items_in_stock and added documentation
Database_Manager.py:
items_in_stock is online and is available for use. It is autogenerated
after each run. This might pose a certain bug for users as it is
functioning on a first run basis. There could be a possibility of adding
administrative priveleges to REPLACE the VIEW for items_in_stock or
create_summary_column. Beliefe is it will be best to leave it as it is
and let the operator decide.
Added documentation for further explanation of each function.

main.py:
Made minor grammatical and astetic modifications.